requirements.txt
# Core LlamaIndex framework (Python 3.13 compatible)
llama-index==0.11.23
llama-index-core==0.11.23
llama-index-llms-ollama==0.3.6
llama-index-embeddings-huggingface==0.3.1
llama-index-vector-stores-chroma==0.2.0

# ML base frameworks (Python 3.13 compatible - using latest)
torch>=2.6.0
transformers>=4.40.0

# Embedding models and vector database
sentence-transformers>=3.2.0
chromadb>=0.5.0  # Updated for Python 3.13 compatibility

# Required dependencies (Python 3.13 compatible)
pytest>=8.0.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
tqdm>=4.66.1
numpy>=1.24.0,<2.0.0  # Compatible with llama-index-core
pandas>=2.0.0,<2.3.0  # Compatible range

# Code quality (Python 3.13 compatible)
flake8>=7.0.0
black>=24.1.0

# Document processing
pdfminer.six>=20221105